Move prefixed composing to record module.
